+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
  - Added support for `export onready var` statement on `class` level
  - Added support for `mastersync` and `puppetsync` functions
+ - Added support for standalone comments in expressions (except for few corner-cases)
  - Added `--diff` option to `gdformat` which acts the same as `--check`, but with additional difference printing
 
 ### Fixed